Convert from CommonJS to ESM
