+++ scripts/poly2req.js
@@ -1,50 +1,48 @@
 #!/usr/bin/env node
 
-'use strict';
+import fs from 'fs';
 
-let fs = require('fs');  // Node 4.x required!
-
-let VERSION = "route_5.0";
+let VERSION = 'route_5.0';
 let SAMPLE_SIZE = 20;
 let NUM_REQUEST = 100;
 let NUM_COORDS = 2;
 let PORT = 5000;
 let url_templates = {
-  "route_5.0": "http://127.0.0.1:{port}/route/v1/driving/{coords}?steps=false&alternatives=false",
-  "route_4.9": "http://127.0.0.1:{port}/viaroute?{coords}&instructions=false&alt=false",
-  "table_5.0": "http://127.0.0.1:{port}/table/v1/driving/{coords}",
-  "table_4.9": "http://127.0.0.1:{port}/table?{coords}"
+  'route_5.0': 'http://127.0.0.1:{port}/route/v1/driving/{coords}?steps=false&alternatives=false',
+  'route_4.9': 'http://127.0.0.1:{port}/viaroute?{coords}&instructions=false&alt=false',
+  'table_5.0': 'http://127.0.0.1:{port}/table/v1/driving/{coords}',
+  'table_4.9': 'http://127.0.0.1:{port}/table?{coords}'
 };
 
 let coord_templates =  {
-  "route_5.0": "{lon},{lat}",
-  "route_4.9": "loc={lat},{lon}",
-  "table_5.0": "{lon},{lat}",
-  "table_4.9": "loc={lat},{lon}"
+  'route_5.0': '{lon},{lat}',
+  'route_4.9': 'loc={lat},{lon}',
+  'table_5.0': '{lon},{lat}',
+  'table_4.9': 'loc={lat},{lon}'
 };
 
 let coords_separators = {
-  "route_5.0": ";",
-  "route_4.9": "&",
-  "table_5.0": ";",
-  "table_4.9": "&"
+  'route_5.0': ';',
+  'route_4.9': '&',
+  'table_5.0': ';',
+  'table_4.9': '&'
 };
-var axis = "distance";
+var axis = 'distance';
 
 var sw = [Number.MAX_VALUE, Number.MAX_VALUE];
 var ne = [Number.MIN_VALUE, Number.MIN_VALUE];
 
-if (process.argv.length > 2 && process.argv[2] == "planet")
+if (process.argv.length > 2 && process.argv[2] == 'planet')
 {
   sw = [-180., -85.];
   ne = [180., 85.];
 }
-if (process.argv.length > 2 && process.argv[2] == "us")
+if (process.argv.length > 2 && process.argv[2] == 'us')
 {
   sw = [-127., 24.];
   ne = [-67., 48.];
 }
-else if (process.argv.length > 2 && process.argv[2] == "dc")
+else if (process.argv.length > 2 && process.argv[2] == 'dc')
 {
   sw = [-77.138, 38.808];
   ne = [-76.909, 38.974];
@@ -54,8 +52,6 @@ else if (process.argv.length > 2)
   let poly_path = process.argv[2];
   let poly_data = fs.readFileSync(poly_path, 'utf-8');
 
-  // lets assume there is only one ring
-  // cut of name and ring number and the two END statements
   let coordinates = poly_data.split('\n')
     .filter((l) => l != '')
     .slice(2, -2).map((coord_line) => coord_line.split(' ')
@@ -78,7 +74,6 @@ if (process.argv.length > 3)
 console.error(sw);
 console.error(ne);
 
-// Yes this an own seeded random number generator because its only a few lines
 var seed = 0x1337;
 function seededRandom(min, max) {
   seed = (seed * 9301 + 49297) % 233280;
@@ -93,11 +88,11 @@ function getRandomCoordinate() {
 }
 
 function makeQuery(coords) {
-  let coords_string = coords.map((c) => coord_templates[VERSION].replace("{lon}", c[0]).replace("{lat}", c[1])).join(coords_separators[VERSION]);
-  return url_templates[VERSION].replace("{coords}", coords_string).replace("{port}", PORT);
+  let coords_string = coords.map((c) => coord_templates[VERSION].replace('{lon}', c[0]).replace('{lat}', c[1])).join(coords_separators[VERSION]);
+  return url_templates[VERSION].replace('{coords}', coords_string).replace('{port}', PORT);
 }
 
-if (axis == "distance")
+if (axis == 'distance')
 {
   for (var i = 0; i < NUM_REQUEST; ++i)
   {
@@ -109,7 +104,7 @@ if (axis == "distance")
     console.log(makeQuery(coords));
   }
 }
-else if (axis == "waypoints")
+else if (axis == 'waypoints')
 {
   for (var power = 0; power <= 1; ++power)
   {
@@ -129,4 +124,3 @@ else if (axis == "waypoints")
     }
   }
 }
-
