+++ test/data/Makefile
@@ -7,9 +7,9 @@ OSRM_CONTRACT:=$(OSRM_BUILD_DIR)/osrm-contract
 OSRM_PARTITION:=$(OSRM_BUILD_DIR)/osrm-partition
 OSRM_CUSTOMIZE:=$(OSRM_BUILD_DIR)/osrm-customize
 OSRM_ROUTED:=$(OSRM_BUILD_DIR)/osrm-routed
-POLY2REQ:=$(SCRIPT_ROOT)/poly2req.cjs
-MD5SUM:=$(SCRIPT_ROOT)/md5sum.cjs
-TIMER:=$(SCRIPT_ROOT)/timer.cjs
+POLY2REQ:=$(SCRIPT_ROOT)/poly2req.js
+MD5SUM:=$(SCRIPT_ROOT)/md5sum.js
+TIMER:=$(SCRIPT_ROOT)/timer.js
 PROFILE:=$(PROFILE_ROOT)/car.lua
 
 all: data
